Move Juz interface to dedicated file
