+++ app/(features)/juz/hooks/useJuzData.ts
@@ -6,17 +6,19 @@ import useVerseListing from '@/app/(features)/surah/hooks/useVerseListing';
 export function useJuzData(juzId?: string) {
   const verseListing = useVerseListing({ id: juzId, lookup: getVersesByJuz });
 
-  const { data: juz, error: juzError } = useSWR<Juz>(juzId ? ['juz', juzId] : null, ([, id]) =>
-    getJuz(id)
+  const { data: juz, error: juzError } = useSWR<Juz>(
+    juzId ? ['juz', juzId] : null,
+    ([, id]: [string, string]) => getJuz(id)
   );
 
-  const isLoading = verseListing.isLoading || (!juz && !juzError);
+  const { isLoading: versesLoading, ...rest } = verseListing;
+  const isLoading = versesLoading || (!juz && !juzError);
 
   return {
     juz,
     juzError,
     isLoading,
-    ...verseListing,
+    ...rest,
   } as const;
 }
