+++ types/index.ts
@@ -5,14 +5,6 @@ export * from './tafsir';
 export * from './settings';
 export * from './surah';
 export * from './word';
+export * from './juz';
 
 export type { TafsirResource } from '../lib/api';
-
-export interface Juz {
-  id: number;
-  juz_number: number;
-  verse_mapping: Record<string, string>;
-  first_verse_id: number;
-  last_verse_id: number;
-  verses_count: number;
-}
